ofproto_v1_5: Function to disable EXT-334
Because OVS 2.9.0 does not yet support the OXS(EXT-334), in order to use
``OFPMP_FLOW_STATS``, ``OFPMP_AGGREGATE_STATS`` and
``OFPT_FLOW_REMOVED`` messages, we need to use parsers for the OpenFlow
version 1.4.

This patch adds an utility function to disable EXT-334 features and to
use the OpenFlow 1.4 parsers for these messages.

Signed-off-by: IWASE Yusuke <iwase.yusuke0@gmail.com>